==> 01_mapper.ipynb <==
import pandas as pd

iupac = pd.read_excel('IUPAC.xlsx')
mapped = pd.read_excel('mapped.xlsx')

iupac.head()

mapped.head()

id_map = pd.Series(iupac['GlyTouCan ID'].values, index = iupac['IUPAC (Derda Lab)'].values)

id_map.head()

name_map = pd.Series(iupac['IUPAC (GlyTouCan)'].values, index = iupac['IUPAC (Derda Lab)'].values)

name_map.head()

mapped['GlyTouCan ID'] = mapped['IUPAC-linker'].map(id_map)

mapped.head()

mapped['GlyTouCan Name'] = mapped['IUPAC-linker'].map(name_map)

mapped

mapped.to_csv('mapped.csv', index = False)

names = pd.read_excel('names.xlsx')

name = names['IUPAC-linker'].unique()

frame = pd.DataFrame()

frame['IUPAC'] = name

frame

frame.to_csv('frame.csv', index = False)
