==> 04_hamming.ipynb <==
# SH-I

import pandas as pd
import numpy as np

def ham_dist(s1, s2):
    
    assert len(s1) == len(s2)
    
    dist = sum(ch1 != ch2 for ch1, ch2 in zip(s1, s2))
    
    return dist

data = pd.read_csv('SS_unfiltered.txt', delim_whitespace = True, skiprows = [1])
data = data.reset_index()

samp = pd.read_csv('SS_unfiltered_sample.txt', delim_whitespace = True, skiprows = [1])
samp = samp.reset_index()
samp = samp.sort_values('SS2', ascending = False)
samp = samp.reset_index(drop = 'True')

samp

x = 1

seq_list1 = list(samp.loc[0])
seq_list2 = list(samp.loc[x])
seq_arr1 = np.array(seq_list1[3:])
seq_arr2 = np.array(seq_list2[3:])
nam_list = seq_list1[:3]
seq_list = list(seq_arr1 + seq_arr2)
fin_list = nam_list + seq_list
samp.loc[0] = fin_list
samp = samp.drop([x])
samp = samp.reset_index(drop = 'True')

samp
