==> 05_array.ipynb <==
# SH-I

import pandas as pd
import numpy as np

def softmax(x):

    e_x = np.exp(x - np.max(x))

    return e_x / e_x.sum()

data = pd.read_csv('SS_unfiltered_output.csv')
data.set_index('Tag', inplace = True)
del data['X']

data = data.sort_values('SS7', ascending = False)

arr = np.array(data['SS7'])
arr = arr.astype(float)
arr.dtype

print(arr)

print(softmax(arr))
