==> 06_overlap.ipynb <==
# SH-I

import pandas as pd
import os
import numpy as np

folder = '20180809-87CLooPA-MR-ppm/'
col_name = 'DNA'

dna_list = []

for filename in os.listdir(folder):
    
    try:
        
        data = pd.read_csv(folder + filename)

        dna = list(data[col_name].values)

        dna_list = dna_list + dna
    
    except:
        
        pass

frame = pd.DataFrame()
frame[col_name] = dna_list
dup = len(dna_list) - len(frame[col_name].unique())

print('Number of duplicates: ' + str(dup))
