==> 07_unique.ipynb <==
# SH-I

import pandas as pd
import numpy as np

input = '20180530-87CLooOO-JM-ppm.txt'

data = pd.read_csv(input, delim_whitespace = True, skiprows = [1])
data.set_index('Nuc', inplace = True)
del data['AA']

for x in range(len(data.columns)):

    data.iloc[x] = data.iloc[x].where(data.iloc[x] == data.iloc[x].max(), 0)
    
    print('Row ' + str(x) + ' fixed.')

data
