==> 09_glytoucan.ipynb <==
# SH-I

import pandas as pd
import numpy as np

data = pd.read_csv('glycans.csv')

excep = ['3S', '4S', '6S', '3P', '4P', '6P', '(Gc)', '-ol', 'Mur', 'Rha']
id_list = []

for x in data['IUPAC (CFG)']:
    
    if any(y in x for y in excep):
        
        id_list.append('inconvertible')
    
    else:
        
        id_list.append('convertible')

data['action'] = id_list

data

data.to_csv('glycans_actions.csv', index = False)
