==> 10_cleaner.ipynb <==
# SH-I

import pandas as pd
import numpy as np

data = pd.read_csv('glytoucan_data.csv')

data = data.dropna()
data = data[~data['IUPAC'].str.contains('Tri|Hex|Hep|Oct|Pen|Non|sugar|dTal|dAlt|Gro|Abe|All|Alt|Ami|Ara|Boi|Asc|Bac|Col|Dig|Ery|Qui')]
data = data.drop_duplicates(subset = ['IUPAC'])

data = data.sort_values('IUPAC')
data = data.reset_index(drop = True)

data

print(str((len(data) / 108978) * 100) + '% of original.')

data.to_csv('glytoucan_data_clean.csv', index = False)
