==> 11_glycan_dict.ipynb <==
# SH-I

import pandas as pd
import numpy as np

data = pd.read_csv('glytoucan_data_clean.csv')
del data['IUPAC (spaced)']
data.set_index('IUPAC', inplace = True)

gly_dict = dict(data['ID'])

gly_dict['Gal(b1-']
