==> 12_conv.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import string

data = pd.read_csv('convertible.csv')

brac = str.maketrans('()', '[]')

def conv(iupac):
    
    iupac = iupac.translate(brac)

    glycan = iupac.split('-')
    
    if len(glycan) == 2:
        
        mono_1 = glycan[0]
        mono_1 = mono_1[:-1] + '(' + mono_1[-1:]
        
        return mono_1 + '1-'

    if len(glycan) == 3:

        mono_1 = glycan[0]
        mono_1 = mono_1[:-2] + '(' + mono_1[-2:]

        mono_2 = glycan[1]
        mono_2 = mono_2[:1] + ')' + mono_2[1:-1] + '(' + mono_2[-1:] + '1'
        
        return mono_1 + '-' + mono_2 + '-'

    else:

        mono_1 = glycan[0]
        mono_1 = mono_1[:-2] + '(' + mono_1[-2:]

        mono_mid = glycan[1:-2]

        for i, x in enumerate(mono_mid):

            x = x[:1] + ')' + x[1:-2] + '(' + x[-2:]

            mono_mid[i] = x

        mono_n = glycan[-2]
        mono_n = mono_n[:1] + ')' + mono_n[1:-1] + '(' + mono_n[-1:] + '1'
        
        return mono_1 + '-' + '-'.join(mono_mid) + '-' + mono_n + '-'

def link(iupac):
    
    glycan = iupac.split('-')
    link = glycan[-1]
    
    return link

def lac(iupac):
    
    iupac = x.iupac('Lac', 'Gal(b1-4)Glc(b1-')
    
    return iupac

data['GTC1'] = data['IUPAC'].apply(conv)
data['GTC2'] = data['GTC1'].str[:-3] + '?1-'
data['Linkage'] = data['IUPAC'].apply(link)

gtc = pd.read_csv('glytoucan_data_clean.csv')
del gtc['IUPAC (spaced)']
gtc.set_index('IUPAC', inplace = True)

gly_dict = dict(gtc['ID'])

data['ID1'] = data['GTC1'].map(gly_dict)
data['ID2'] = data['GTC2'].map(gly_dict)
data['ID'] = data['ID1'].fillna(data['ID2'])

del data['ID1']
del data['ID2']

data

data.to_csv('test.csv', index = False)
