==> 14_array_maker.ipynb <==
# SH-I

import pandas as pd
import numpy as np

filename = '20180222-87CLooPA-MR-ppm'
data = pd.read_csv(filename + '.txt', delim_whitespace = True)
data.set_index('Nuc', inplace = True)

def ham_dist(s1, s2):

    assert len(s1) == len(s2)

    dist = sum(ch1 != ch2 for ch1, ch2 in zip(s1, s2))

    return dist

col_df = data['1-1'].sort_values(ascending = False)
col_df = col_df[col_df > 0]

col_df

n_list = list(range(len(col_df) - 1))

for n in n_list:

    seq1 = col_df.index[n]
    
    print('Checking:')
    print(seq1)

    count = list(range(n + 1, len(col_df)))

    for x in count:

        seq2 = col_df.index[x]

        try:

            dist = ham_dist(seq1, seq2)

        except AssertionError:

            dist = None

        if dist == 0 or dist == 1:

            col_df[n] = col_df[n] + col_df[x]
            col_df = col_df.drop([seq2])

            del count[-1]
            del n_list[-1]

col_df
