==> 18_intersection.ipynb <==
# SH-I

import pandas as pd
import numpy as np

cfg = pd.read_excel('IUPAC (CFG).xlsx')
liga = pd.read_excel('IUPAC (LiGA).xlsx')

cfg_gly = cfg['IUPAC (GlyTouCan)'].values
liga_gly = liga['IUPAC (GlyTouCan)'].values

output = pd.merge(cfg, liga, on = 'IUPAC (GlyTouCan)')

output.to_csv('intersection.csv', index = False)
