==> 19_maldi.ipynb <==
# SH-I

import pandas as pd
import numpy as np

maldi = pd.read_excel('maldi.xlsx')
yg = pd.read_excel('yg.xlsx')

maldi_arr = maldi['IUPAC (MALDI)'].unique()
yg_arr = yg['IUPAC (YG)'].unique()

maldi
