==> 20_yg.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import os

cfg = pd.read_excel('IUPAC (CFG).xlsx')
yg = pd.read_excel('IUPAC (YG).xlsx')

del cfg['Linkage (CFG)']
del yg['Linakge']

data = pd.merge(cfg, yg, on = 'IUPAC (GlyTouCan)')
data = data.sort_values('IUPAC (GlyTouCan)')
data['Structure on Masterlist'] = data['IUPAC (CFG)']

for filename in os.listdir('workable/'):
    
    file = pd.read_excel('workable/' + filename)
    file = file[['Chart Number', 'Structure on Masterlist', 'Average RFU', 'StDev', '% CV']]
    
    merged = pd.merge(file, data, on = 'Structure on Masterlist')
    
    output = pd.DataFrame()
    output = merged[['Structure on Masterlist', 'Average RFU', 'StDev', '% CV']]
    
    output.to_csv(filename[:-4] + '.csv', index = False)
