==> 21_norm.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

data = pd.read_csv('lectins_rfu.csv')
data.set_index('IUPAC', inplace = True)

def norm(arr):
    
    p = (arr - arr.min())
    q = arr.max() - arr.min()
    
    return p / q

for col_name in data.columns:
    
    arr = data[col_name].values
    arr_n = norm(arr)
    data[col_name] = arr_n

def vectorized_result(i):
    
    upper = 0.6
    lower = 0.2

    if upper <= i:

        val = 'high'

    elif lower <= i < upper:

        val = 'med'

    elif i < lower:

        val = 'low'

    return val

data = data.applymap(vectorized_result)

data.to_csv('lectins_val.csv')
