==> 22_target.ipynb <==
# SH-I

import pandas as pd
import numpy as np

data = pd.read_csv('rfu.csv')
data.set_index('Structure on Masterlist', inplace = True)

col = np.array(data.columns)
col = np.sort(np.sort(col))

def tar(col_name):
    
    tar_list = col_name.split('_')
    tar_id = tar_list[-1]
    tar = '_'.join(tar_list[:-1])
    
    return tar, tar_id

x_list = []
y_list = []

for col_name in col:
    
    x, y = tar(col_name)
    
    x_list.append(x)
    y_list.append(y)
    
    print(y)

tar = pd.DataFrame()
tar['File Name'] = x_list
tar['File ID'] = y_list

tar.to_csv('tar.csv', index = False)
