==> 24_names.ipynb <==
# SH-I

import pandas as pd

tar = pd.read_csv('tar.csv')
tar = tar.sort_values('File ID')
tar = tar.reset_index(drop = True)

tar_list = []

for i, name in enumerate(tar['Target']):
    
    if ' Slide#' in name:
        
        name_list = name.split(' Slide#')
    
    elif ' slide#' in name:
        
        name_list = name.split(' slide#')
    
    if ' CFG#' in name_list[0]:
        
        name_list = name_list[0].split(' CFG#')
    
    print(i, name_list[0])
    
    tar_list.append(name_list[0])

n = 649

print(tar.iloc[n, 2])

tar['Target Name'] = tar_list

tar.to_csv('tar.csv', index = False)
