==> 26_maldi_iupac.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import converter as cn

data = pd.read_excel('MALDI-IUPAC.xlsx')

data['IUPAC'] = data['IUPAC (LiGA)'].apply(cn.conv).apply(cn.lac)
data['Linkage'] = data['IUPAC (LiGA)'].apply(cn.link)

data

data.to_csv('MALDI-IUPAC.csv', index = False)
