==> 27_maldi_id.ipynb <==
# SH-I

import pandas as pd
import numpy as np

iupac = pd.read_excel('IUPAC.xlsx')
maldi = pd.read_csv('MALDI-IUPAC.csv')

maldi

id_dict = dict(zip(iupac['IUPAC (GlyTouCan)'], iupac['GlyTouCan ID']))

maldi['GlyTouCan ID'] = maldi['IUPAC'].map(id_dict)

maldi

maldi.to_csv('MALDI-IUPAC.csv')
