==> 28_maldi_map.ipynb <==
# SH-I

import pandas as pd
import numpy as np

maldi_map = pd.read_excel('malid-map.xlsx')
maldi = pd.read_csv('MALDI-IUPAC.csv')

iupac_dict = dict(zip(maldi['IUPAC (LiGA) Non-repeating'], maldi['IUPAC']))
link_dict = dict(zip(maldi['IUPAC (LiGA) Non-repeating'], maldi['Linkage']))
id_dict = dict(zip(maldi['IUPAC (LiGA) Non-repeating'], maldi['GlyTouCan ID']))

maldi_map['IUPAC'] = maldi_map['IUPAC Linker'].map(iupac_dict)
maldi_map['Linkage'] = maldi_map['IUPAC Linker'].map(link_dict)
maldi_map['GlyTouCan ID'] = maldi_map['IUPAC Linker'].map(id_dict)
maldi_map

maldi_map.to_csv('malid-map.csv', index = False)
