==> 29_peptide_seq1.ipynb <==
# SH-I

import pandas as pd
import numpy as np

from urllib.request import urlopen

tar = pd.read_csv('tar.csv')

def psId(term):
    
    html = urlopen('http://www.functionalglycomics.org/glycomics/search/jsp/result.jsp?query=' + term + '&cat=all')

    page = str(html.read())
    page_list = page.split('/glycomics/HServlet?operation=view&sideMenu=no&psId=')[1]
    
    return page_list.split('"')[0]

ps_list = []

for term in tar['Target Name']:
    
    if 'ug/ml)' in term:
        
        name = term.split(' (')[0]
    
    elif 'Dilution)' in term:
        
        name = term.split(' (')[0]
    
    elif 'ug/ml' in term:
        
        name = term.split(' ')[:-1]
        name = ' '.join(name)
    
    else:
        
        name = term
    
    try:
        
        psid = psId(name)

        print(psid)
        
        ps_list.append(psid)
    
    except:
        
        print(name)
        
        ps_list.append(None)

tar['primscreen ID'] = ps_list

tar.to_csv('tar.csv', index = False)
