==> 32_recluster.ipynb <==
# SH-I

import pandas as pd
import numpy as np

yo = pd.read_csv('re_YO_conv.csv')
uea = pd.read_csv('focus_folder/Ulex europaeus Agglutinin.csv')

# glycan name convertions to IUPAC
yo_conv = pd.read_csv('focus_folder/MALDI-IUPAC.csv')
uea_conv = pd.read_excel('focus_folder/convertible.xlsx')

# dictionaries to IUPAC
uea_dict = dict(zip(uea_conv['IUPAC'], uea_conv['GlyTouCan'] + uea_conv['Linkage']))

uea['Structure_IUPAC'] = uea['Structure on Masterlist'].map(uea_dict)
uea = uea[['Structure_IUPAC', 'Average RFU', 'StDev', '% CV']]

yo_gly = yo['YO (IUPAC)']
uea_gly = uea['Structure_IUPAC']

yo_name = []
yo_link = []

for i in yo_gly:
    
    x = i.split('-')[:-1]
    x = '-'.join(x) + '-'
    
    y = i.split('-')[-1]
    
    yo_name.append(x)
    yo_link.append(y)

uea_name = []
uea_link = []

for i in uea_gly:
    
    x = i.split('-')[:-1]
    x = '-'.join(x) + '-'
    
    y = i.split('-')[-1]
    
    uea_name.append(x)
    uea_link.append(y)

yo_arr = np.array(yo_name)
uea_arr = np.array(uea_name)

yo_df = pd.DataFrame()
yo_df['Name (YO)'] = yo_name

uea_dict = dict(zip(uea_name, uea_gly))

yo_df['Glycans (CFG)'] = yo_df['Name (YO)'].map(uea_dict)

arr = np.setdiff1d(uea['Structure_IUPAC'], yo_df['Glycans (CFG)'].fillna(''))

df = pd.DataFrame()
df['Glycans'] = arr

df.to_csv('df.csv', index = False)
