==> 35_peptide_seq4.ipynb <==
# SH-I

import pandas as pd
import numpy as np

from urllib.request import urlopen

tar = pd.read_csv('tar.csv')
tar = tar.fillna(' ')

def psId(term):

    html = urlopen('http://www.functionalglycomics.org/glycomics/search/jsp/result.jsp?query=' + term + '&cat=all')

    page = str(html.read())
    page_list = page.split('/glycomics/HServlet?operation=view&sideMenu=no&psId=')[1]

    return page_list.split('"')[0]

def cbpId(psid):

    html = urlopen('http://www.functionalglycomics.org/glycomics/HServlet?operation=view&sideMenu=no&psId=' + psid)
    page = str(html.read())

    cbpid = page.split('cbpId=')[1]
    cbpid = cbpid.split('&sideMenu=no')[0]
    
    return cbpid

def seq(cbpid):

    html = urlopen('http://www.functionalglycomics.org/glycomics/GBPServlet?pageType=view&view=view&operationType=view&cbpId=' + cbpid + '&sideMenu=no')
    page = str(html.read())

    seq = page.split('align="left" class="rightSideEqualWidth">')[1]
    seq = seq.split('</td>')[0]
    seq = ' '.join(seq.split('&nbsp;'))
    seq = ' '.join(seq.split('<br>'))

    return seq

for i in list(tar.index):
    
    tar_name = tar.loc[i, 'Target Name']
    psid = tar.loc[i, 'primscreen ID']
    cbpid = tar.loc[i, 'CBP ID']
    seq = tar.loc[i, 'Sequence']
    
    if psid == ' ':
        
        if 'ug/ml' in tar_name:
            
            name = tar_name.split(' ')[:-1]
            name = ' '.join(name)
            
            print(name)
            print('')
            
        elif 'Dilution' in tar_name:
            
            name = tar_name.split('(')[:-1]
            name = ' '.join(name)
            
            print(name)
            print('')
        
        else:
            
            print(tar_name)
            print('')

tar
