==> 40_vectorized.ipynb <==
# SH-I

import numpy as np

def vectorized_result(i):
    
    e = np.zeros((3, 1))
    
    if 0.66 <= i:
        
        e[0] = 1.0
    
    elif 0.33 <= i < 0.66:
        
        e[1] = 1.0
    
    elif i < 0.33:
        
        e[2] = 1.0
    
    return e

vectorized_result(0.75)
