==> 43_alignment.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

cfg = pd.read_csv('cfg_fig/cfg.csv')
liga = pd.read_csv('cfg_fig/liga_uea.csv')

cfg['Unit'] = ['-'.join(x.split('-')[:-1]) + '-' for x in cfg['IUPAC']]
liga['Unit'] = ['-'.join(x.split('-')[:-1]) + '-' for x in liga['IUPAC']]

df = pd.merge(liga, cfg, on = 'Unit', how = 'outer')
df = df[['Mod', 'total', 'IUPAC_x', 'IUPAC_y', 'Structure on Masterlist', 'Average RFU', 'StDev']]
df = df.drop_duplicates()

is_duplicate = df[['Mod', 'total', 'IUPAC_x']].duplicated()
df_liga = df[['Mod', 'total', 'IUPAC_x']].where(~is_duplicate, np.NaN)

is_duplicate = df[['IUPAC_y', 'Structure on Masterlist', 'Average RFU', 'StDev']].duplicated()
df_cfg = df[['IUPAC_y', 'Structure on Masterlist', 'Average RFU', 'StDev']].where(~is_duplicate, np.NaN)

out = df_liga.join(df_cfg)
out = out.drop_duplicates()
out = out.dropna(how = 'all')

out.to_csv('cfg_fig/aligned.csv', index = False)
