==> 50_extension.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import re
import itertools

test = ['ABC-Sp', # linear
        'A[B]C-Sp', # single branch
        'A[B]C[D]E-Sp', # multiple branches
        'A[B]C[D]E[F]G-Sp',
        'A[B]C[D]E[F]G[H]I-Sp', 
        'A[B]C[D[E]F]G-Sp', # sub branches
        'A[B]C[D]E[F[G]H]I-Sp',
        'A[B][C]D-Sp'] # triple braches

def seperate(family):
    
    parent = family.split(']')[-1]
    children = ']'.join(family.split(']')[:-1])
    child = children.split('[')

    for i, x in enumerate(child):

        if ']' not in x:

            pos = i

    child1 = '['.join(child[:pos])
    child2 = '['.join(child[pos:])
    
    return child1, child2, parent

glycan = test[5]
glycan_list = [glycan]

family = '-'.join(glycan.split('-')[:-1])
linkage = glycan.split('-')[-1]

child1, child2, parent = seperate(family)

glycan_list.append(child2 + '[' + child1 + ']' + parent + '-' + linkage)

glycan_list

seperate(family)

for x in test:
    
    print(x)
