==> 52_stratified.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import itertools

def load_data(protein, split):

    i1, i2 = int((5 * split) / 2), 5

    data = pd.read_csv('data/lectins_norm.csv')
    data = data.round(decimals = 4)
    data = data.sort_values(protein, ascending = False)

    input_data = list(data['IUPAC'])
    output_data = list(data[protein])

    if split == 0:

        training_data = zip(input_data, output_data)
        test_data = zip(input_data, output_data)

    else:

        input_strata = input_data[:i1:i2] + input_data[-i1::i2]
        output_strata = output_data[:i1:i2] + output_data[-i1::i2]

        training_data = zip([i for i in input_data if not i in input_strata or input_strata.remove(i)],
                            [i for i in output_data if not i in output_strata or output_strata.remove(i)])

        test_data = zip(input_data[:i1:i2] + input_data[-i1::i2],
                        output_data[:i1:i2] + output_data[-i1::i2])

    return (training_data, test_data)

def load_encoded(protein, split = 10):

    training_data, test_data = load_data(protein, split)

    training_input, training_output = [], []
    test_input, test_output = [], []

    for x, y in training_data:

        training_input.append(encoder(x))
        training_output.append(vectorized_result(y))

    for x, y in test_data:

        test_input.append(encoder(x))
        test_output.append(vectorized_result(y))

    training_encoded = zip(training_input, training_output)
    test_encoded = zip(test_input, test_output)

    return (training_encoded, test_encoded)

def save_loaded(protein, split):

    training_data, test_data = load_data(protein, split)

    x_list = []
    y_list = []

    for x, y in test_data:

        x_list.append(x)
        y_list.append(y)

    test = pd.DataFrame()

    test['x'] = x_list
    test['y'] = y_list
    
    test.to_csv('test.csv', index = False)

save_loaded('UEA', 100)
