==> 56_merge_names.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

lectin = 'UEA'

aligned = pd.read_csv('figures/outputs/' + lectin + '.csv')

def merge(aligned):
    
    liga = aligned['IUPAC (LiGA)']
    cfg = aligned['IUPAC (CFG)']

    mod_iupac = []

    for l, c in zip(liga, cfg):

        if c is np.NaN:

            link = '-' + str(l).split('-')[-1]
            glycan = '-'.join(str(l).split('-')[:-1])

        elif l is np.NaN:

            link = '-' + str(c).split('-')[-1]
            glycan = '-'.join(str(c).split('-')[:-1])

        else:

            l_link = str(l).split('-')[-1]
            c_link = str(c).split('-')[-1]

            link = '-' + l_link + '/' + c_link
            glycan = '-'.join(str(l).split('-')[:-1])
            
        mod = glycan + link
        
        if len(mod) > 43:
            
            mod = mod[:30] + '...' + mod[-10:]

        mod_iupac.append(mod)

    aligned['Mod IUPAC'] = mod_iupac
    
    return aligned

dataframe = merge(aligned)

for x in dataframe['Mod IUPAC']:
    
    print(x, len(x))
