==> 60_human.ipynb <==
# SH-I

import os
import pandas as pd

lectins = ['ACG', 'CTB', 'DC_SIGN', 'Gal4', 'LCA', 'MAL', 'NPL', 'PSA', 'SBA', 'SNL', 'UEA']

def vectorized_result(i):

    upper = 0.6
    lower = 0.2

    if upper <= i:

        bind = 'high'

    elif lower <= i < upper:

        bind = 'med'

    elif i < lower:

        bind = 'low'

    return bind

def copy(iupac, val):
    
    draw_dir = os.getcwd() + '/drawings/png/cfg/'
    bind_dir = os.getcwd() + '/human/' + lectin + '/'
    
    for filename in os.listdir(draw_dir):

        if iupac in filename:
                
            draw = draw_dir + filename
            bind = bind_dir + val + '/' + filename
                
            os.system('cp "' + draw + '" "' + bind + '"')

def make(lectin):
    
    data = pd.read_csv('data/lectins_norm.csv')

    for bind in ['high', 'med', 'low']:

        if not os.path.exists('human/' + lectin + '/' + bind):

            os.makedirs('human/' + lectin + '/' + bind)
            
    for iupac, val in zip(data['IUPAC'], data[lectin].apply(vectorized_result)):
    
        copy(iupac, val)

for lectin in lectins:
    
    make(lectin)
