==> 65_figures_drawing.ipynb <==
# SH-I

import pandas as pd
import numpy as np
import os
import matplotlib as mpl
import matplotlib.pyplot as plt

lectins = ['ACG', 'CTB', 'DC_SIGN', 'Gal4', 'LCA', 'MAL', 'NPL', 'PSA', 'SBA', 'SNL', 'UEA']

lectin_dict = {'ACG' : 'Agrocybe cylindracea lectin',
               'CTB' : 'Cholera Toxin B subunit',
               'DC_SIGN' : 'DC-SIGN CRD',
               'Gal4' : 'Gal4-mAb',
               'LCA' : 'Lens culinaris Agglutinin',
               'MAL' : 'Maachkia amurensis Leukoagglutinin',
               'NPL' : 'Narcissus pseudonarcissus lectin',
               'PSA' : 'Pisum sativum Agglutinin',
               'SBA' : 'Soyabean Agglutinin',
               'SNL' : 'Sambucus nigra lectin',
               'UEA' : 'Ulex europaeus Agglutinin'}

def numbered(aligned):

    cfg = aligned['IUPAC (CFG)']

    cfg_count = 1
    cfg_count_list = []

    for x in cfg:

        if x is not np.NaN:

            cfg_count_list.append(str(cfg_count))

            cfg_count += 1

        else:

            cfg_count_list.append('')

    aligned['CFG Count'] = cfg_count_list
        
    return aligned

def drawer(aligned, lectin):

    for c, n in zip(aligned['CFG Count'], aligned['IUPAC (CFG)']):

        for filename in os.listdir(os.getcwd() + '/drawings/eps/cfg'):

            if n in filename:

                os.system('cp "' + os.getcwd() + '/drawings/eps/cfg/' + filename +
                            '" "' + os.getcwd() + '/CFG_drawings_hits/' + lectin + '_' + c + '_' + n + '.eps"')

lectin = 'UEA'

data = pd.read_csv('CFG_data_hits/' + lectin + '_top.csv')

drawer(numbered(data), lectin)

for lectin in lectins:
    
    data = pd.read_csv('CFG_data_hits/' + lectin + '_top.csv')
    drawer(numbered(data), lectin)
